generalize push to work for non-dask chunked arrays
